refactor(init): give the source scanner a home and collapse the server scaffolders
Quality pass over the scaffolding changes on this branch.

- extract maskCommentsAndStrings + findMatchingDelimiter into source-scan.ts.
  The masker is the subtlest code here — it carries an index-alignment
  invariant three modules depend on — and it was sitting in a grab-bag file
  next to Tailwind header markup. expo.ts also kept a private copy of the
  delimiter scanner; there is now one implementation.
- replace needsManualWiring, which drove control flow by matching the prose
  of a skip reason ("Could not find…"), with an explicit `wired` flag
  returned by scaffoldServerEntry. Rewording a user-facing message can no
  longer silently drop a setup instruction.
- move plan assembly into scaffoldServerFramework so express.ts and
  fastify.ts stop duplicating it. The quickstart URL and manual-wiring
  sentence are now config, so fastify.ts is pure configuration and
  express.ts only adds its types/globals.d.ts action.
- findStatementEnd takes masked source like the other scanners instead of
  re-masking content its caller had already masked.

No behavior change: 2217 tests pass, and the real create-expo-app project
still scaffolds to identical output.
